Fix: 일부버그 수정
코드저장 DB범위를 10000바이트로 늘림
score변수 double로 변경

+++ src/main/java/com/tools/codemos/dto/CodeDTO.java
@@ -7,10 +7,9 @@ import lombok.Setter;
 @Setter
 public class CodeDTO {
     private int id;
-    private int score;
+    private double score;
     private String code;
 
-    public CodeDTO(int id, int score, String code) {
+    public CodeDTO(int id, double score, String code) {
     }
-
 }
